Hit Analizer Improvements
Added hit player name again:
Knowing who we are shooting at could be taken as an advantage,
but most servers support the /an plugin, so it should be safe to bring this back.

Added option to toggle time since last hit:
IIRC this were used for multibullet detection, loses its purpose as a local player.
However, could still be useful for debugging, so i'll keep it with cg_hitAnalizer 2.